requirements.txt
langchain==0.2.14
langchain_community==0.2.12
langchain_core==0.2.33
langchain_google_vertexai==1.0.8
langchain_groq==0.1.9
lida==0.0.14
matplotlib==3.9.1.post1
pandas==2.2.2
python-dotenv==1.0.1
streamlit==1.38.0
